Change the controller to 2 joysticks.
Change the controller to 2 joysticks, add a emergency button,
fix some socket api, now will attempt to reconnect.

+++ app/src/main/java/com/MobileAnarchy/Android/Widgets/Joystick/JoystickMovedListener.java
@@ -1,6 +1,6 @@
 package com.MobileAnarchy.Android.Widgets.Joystick;
 
 public interface JoystickMovedListener {
-    void OnMoved(int pan, int tilt);
+    void OnMoved(double pan, double tilt);
     void OnReleased();
 }
